image.h: .pnm: better comment parsing
For PNM files (and potentially for the other textual file formats if I'd like to) comments are now being preprocessed out instead of dynamically being ignored during execution. It stabilizes the execution of the parser significantly with the potential cost of time complexity (necessity of processing the same string 2 times, first to get the preprecessed size and the second to copy the it to the new buffer).