image.h: .png: first attempt at decoding
This is the first attempt at decoding the PNG IDAT chunk. It was made thanks to grokai and zlib library. It kind-of works: data is properly loaded but (at least at the first glance) incorrectly processed (tip: save the loaded PNG as a PPM file). Probably something off with ordering.